Remove non-standard values for `zoom` https://bugs.webkit.org/show_bug.cgi?id=280557 rdar://136866582
Reviewed by Tim Nguyen.

This patch is to align WebKit with web specification [1]:

[1] https://drafts.csswg.org/css-viewport/#zoom-property

The `zoom` property was non-standard previously but have been standardised now.
In newer standard, it does not have `document` and `reset` values.
So this patch is to remove them and Blink actually removed them in 2017.

* Source/WebCore/css/CSSProperties.json:
* Source/WebCore/css/CSSValueKeywords.in:
* Source/WebCore/style/StyleBuilderCustom.h:
(WebCore::Style::BuilderCustom::applyValueZoom):
* LayoutTests/imported/w3c/web-platform-tests/css/css-viewport/zoom/parsing/zoom-valid-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/css-viewport/zoom/parsing/zoom-valid.html:

Canonical link: https://commits.webkit.org/284427@main