`-[WKWebViewConfiguration writingToolsBehavior]` not available when using a deployment target lower than iOS 18 https://bugs.webkit.org/show_bug.cgi?id=280645 rdar://136830527
Reviewed by Richard Robinson and Abrar Rahman Protyasha.

The property is unavailable when using a deployment target lower than iOS 18,
even when wrapping code in a `@available(iOS 18.1, *)` block.

This is due to incorrect use of the `MIN_REQUIRED` macro, which is based on the
deployment target. Consequently, the API is not found in the header, even though
the code in the `@available` block will only run on iOS 18.1+.

Fix by using `MAX_ALLOWED` which accounts for the actual SDK version.

* Source/WebKit/UIProcess/API/Cocoa/WKWebViewConfiguration.h:

Canonical link: https://commits.webkit.org/284478@main