Build broken with LLVM 19 https://bugs.webkit.org/show_bug.cgi?id=280642
Reviewed by NOBODY (OOPS!).

We need to replace several forward declarations with #include
statements to avoid "member access into incomplete type" errors. This
has revealed that many SVG headers are missing from the framework
headers and so are not available to derived sources. I don't attempt to
fix this comprehensively; this is just the minimum required to fix the
build.

* Source/WebCore/CMakeLists.txt:
* Source/WebCore/Headers.cmake:
* Source/WebCore/Modules/webcodecs/WebCodecsVideoFrame.h: