Grab fractional scale factor from GSK directly in WebKitGTK. https://bugs.webkit.org/show_bug.cgi?id=271714
Reviewed by NOBODY (OOPS!).

With this change, WebKitGTK doesn't render at the next integer scale (e.g. 150% -> 200%)
only for GTK to scale it down bilinearly afterwards (resulting in a blurry image)
but instead renders at the correct resolution from the beginning.
The widget output may still be a little blurry when the WebView isn't perfectly aligned
to the pixel grid, but this was already the case beforehand (and screenshots look perfect).

When using the Skia renderer on MotionMark, I noticed that some tests are rendered
at the lower logical resolution, but everything looks fine on Cairo.

* Source/WebKit/UIProcess/API/gtk/WebKitWebViewBase.cpp:
(deviceScaleFactorChanged):
(webkitWebViewBaseRealize):
(webkitWebViewBaseUnrealize):
(webkitWebViewBaseCreateWebPage):