[WPE][Debug] ASSERTION FAILED: canSafelyBeUsed() https://bugs.webkit.org/show_bug.cgi?id=280617
Reviewed by Chris Dumez.

Save the host file descriptor on AcceleratedSurfaceLibWPE constructor
that is called in the main thread. In initialize(), called from the
compositor thread when the GL context is created, we can release the
file descriptor passing the ownership to the backend.

* Source/WebKit/WebProcess/WebPage/WebPage.h:
* Source/WebKit/WebProcess/WebPage/libwpe/AcceleratedSurfaceLibWPE.cpp:
(WebKit::AcceleratedSurfaceLibWPE::AcceleratedSurfaceLibWPE):
(WebKit::AcceleratedSurfaceLibWPE::initialize):
* Source/WebKit/WebProcess/WebPage/libwpe/AcceleratedSurfaceLibWPE.h:

Canonical link: https://commits.webkit.org/284554@main