[Static Analyzer CI] Add steps to download cmake and ninja on worker https://bugs.webkit.org/show_bug.cgi?id=280722 rdar://137073482
Reviewed by Jonathan Bedard.

In order to build clang, we need to install cmake and ninja. This commit introduces a script
that downloads the corresponding package and installs it on the worker in a specified directory.

The InstallClang and InstallNinja steps are added to the factory and should only need to be run
on a bot the very first time it is running this particular queue.

* Tools/CISupport/ews-build/factories.py:
(SmartPointerStaticAnalyzerFactory.__init__):
* Tools/CISupport/ews-build/steps.py:
(InstallCMake):
(InstallCMake.__init__):
(InstallCMake.run):
(InstallCMake.evaluateCommand):
(InstallCMake.getResultSummary):
(InstallNinja):
(InstallNinja.__init__):
(InstallNinja.run):
(InstallNinja.evaluateCommand):
(InstallNinja.getResultSummary):
* Tools/CISupport/ews-build/steps_unittest.py:

Canonical link: https://commits.webkit.org/284572@main