REGRESSION (279761@main): Returning to inline playback from fullscreen shows only the play/pause button instead of the full controls https://bugs.webkit.org/show_bug.cgi?id=280903 rdar://134924988
Reviewed by Andy Estes.

Full control should be displayed when switching back from fullscreen mode, but play/pause button is displayed.

Steps to reproduce:

1. Open https://developer.apple.com/videos/play/wwdc2023/10279/ in Safari on macOS
2. Start video playback
3. Enter fullscreen
4. Exit fullscreen

At this point you’ll notice that the only media control available is the play/pause button instead of the full media controls.

The issue is introduced with the following commit:
https://commits.webkit.org/279761@main

The logic added to HTMLMediaElement::shouldForceControlsDisplay() is for tvOS and no reason to apply it to other platform.
Wrapping the logic with PLATFORM(APPLETV) should fix the issue.

* Source/WebCore/html/HTMLMediaElement.cpp:
(WebCore::HTMLMediaElement::shouldForceControlsDisplay const):

Canonical link: https://commits.webkit.org/284711@main