Avoid process suspension on launch https://bugs.webkit.org/show_bug.cgi?id=280969 rdar://137418855
Reviewed by NOBODY (OOPS!).

Currently, the first assertion we take on process launch is the boosted jetsam assertion. This assertion
does not provide any runtime, so the system will immediately suspend the process when this assertion is
taken. To avoid this suspension, we can take a foreground assertion via the process throttler before we
take the jetsam assertion. The foreground assertion is taken async when the process throttler has been
connected to the process.

* Source/WebKit/UIProcess/AuxiliaryProcessProxy.cpp:
(WebKit::AuxiliaryProcessProxy::didFinishLaunching):