RELEASE_ASSERT that ImageIO could install all the supported image decoders https://bugs.webkit.org/show_bug.cgi?id=280990 rdar://135375243
Reviewed by Simon Fraser.

We needs to check the return value of CGImageSourceSetAllowableTypes().

If CGImageSourceSetAllowableTypes() could not install all the supported image
decoders for any reason, we should halt the process launch and fail its creation
safely.

Because the WebPage can be reinitialized (WTR for example), it is important to
call GImageSourceSetAllowableTypes() only once and only when a WebProcess is
initialized.

* Source/WebCore/platform/graphics/cg/UTIRegistry.mm:
(WebCore::defaultSupportedImageTypes):
* Source/WebCore/platform/network/mac/UTIUtilities.mm:
(WebCore::setImageSourceAllowableTypes):

Canonical link: https://commits.webkit.org/284850@main