Don't repaint SVG elements not in tree and check for nullptr before derefencing enclosing layer https://bugs.webkit.org/show_bug.cgi?id=281243 rdar://137178583 ([bddb15f89f40df55] ASAN_SEGV | WebCore::RenderElement::isInsideEntirelyHiddenLayer; WebCore::SVGRenderSupport::clippedOverflowRectForRepaint; WebCore::RenderObject::issueRepaint)
Reviewed by NOBODY (OOPS!).

When we repaint SVG elements we check if they are on the
document view before issuing the repaint, also if the
enclosingLayer of a RenderElement is null, we don't check
if it has visible content.
This avoid the null pointer deref.

* LayoutTests/fast/svg/render-element-no-enclosing-layer-expected.txt: Added.
* LayoutTests/fast/svg/render-element-no-enclosing-layer.html: Added.
* Source/WebCore/rendering/ReferencedSVGResources.cpp:
(WebCore::CSSSVGResourceElementClient::resourceChanged):
* Source/WebCore/rendering/RenderElement.cpp:
(WebCore::RenderElement::isInsideEntirelyHiddenLayer const):
(WebCore::RenderElement::repaintOldAndNewPositionsForSVGRenderer const):